Move 'URLProtocolStub' to its own file
