+++ MenuAppTDD/MenuAppTDDTests/URLProtocolStub.swift
@@ -1,0 +1,72 @@
+//
+//  URLProtocolStub.swift
+//  MenuAppTDDTests
+//
+//  Created by Vinicius Moreira Leal on 19/05/2020.
+//  Copyright © 2020 Vinicius Leal and Gabriel Meira. All rights reserved.
+//
+
+import Foundation
+
+// This class will allow us to pass Data, URLResponse, and Error, to simulate the behaviour of a network request.
+class URLProtocolStub: URLProtocol {
+    private static var stub: Stub?
+    private static var requestObserver: ((URLRequest) -> Void)?
+
+    private struct Stub {
+        let data: Data?
+        let response: URLResponse?
+        let error: Error?
+    }
+
+    enum TestFailureCondition: Error {
+        case invalidServerResponse
+    }
+
+    static func stub(data: Data?, response: URLResponse?, error: Error?) {
+        stub = Stub(data: data, response: response, error: error)
+    }
+
+    static func observeRequests(observer: @escaping (URLRequest) -> Void) {
+        requestObserver = observer
+    }
+
+    static func startInterceptingRequests() {
+        URLProtocol.registerClass(URLProtocolStub.self)
+    }
+
+    static func stopInterceptingRequests() {
+        URLProtocol.unregisterClass(URLProtocolStub.self)
+        stub = nil
+        requestObserver = nil
+    }
+
+    override class func canInit(with request: URLRequest) -> Bool { true }
+
+    override class func canonicalRequest(for request: URLRequest) -> URLRequest {
+        return request
+    }
+
+    override func startLoading() {
+        if let requestObserver = URLProtocolStub.requestObserver {
+            client?.urlProtocolDidFinishLoading(self)
+            return requestObserver(request)
+        }
+
+        if let data = URLProtocolStub.stub?.data {
+            client?.urlProtocol(self, didLoad: data)
+        }
+
+        if let response = URLProtocolStub.stub?.response {
+            client?.urlProtocol(self, didReceive: response, cacheStoragePolicy: .notAllowed)
+        }
+
+        if let error = URLProtocolStub.stub?.error {
+            client?.urlProtocol(self, didFailWithError: error)
+        }
+
+        client?.urlProtocolDidFinishLoading(self)
+    }
+
+    override func stopLoading() {}
+}
